Add hello world task and restructured tasks
