+++ typescript-tasks/001-hello-world/001-hello-world.spec.ts
@@ -1,0 +1,8 @@
+import { helloWorld } from "./001-hello-world";
+
+describe("helloWorld", () => {
+  it("Should return string of Hello World", async () => {
+    const result = helloWorld();
+    expect(result).toEqual("Hello World");
+  });
+});
