+++ typescript-tasks/001-hello-world/001-hello-world.ts
@@ -1,0 +1,6 @@
+/* NAME:
+ * Hello World
+ *
+ * TASK:
+ * Create a function that returns a string "Hello World".
+ */
