+++ typescript-tasks/003-fizzbuzz-switch-statement/003-fizzbuzz-switch-statement.ts
