+++ typescript-tasks/004-fizzbuzz-chain-responsibility/004-fizzbuzz-chain-responsibility.ts
